@@ -2,7 +2,7 @@ package com.fdkservice.security.payload.request;
 
 import java.util.Set;
 
-import javax.validation.constraints.*;
+import jakarta.validation.constraints.*;
  
 public class SignupRequest {
     @NotBlank
