@@ -1,11 +1,10 @@
-package com.fdkservice.security.security.services;
+package com.fdkservice.security.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.fdkservice.security.models.User;
 import com.fdkservice.security.repository.UserRepository;
@@ -16,7 +15,6 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 	UserRepository userRepository;
 
 	@Override
-	@Transactional
 	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 		User user = userRepository.findByUsername(username)
 				.orElseThrow(() -> new UsernameNotFoundException("User Not Found with username: " + username));
